app/requirements.txt
atlassian-python-api==3.28.1
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
idna==3.4
oauthlib==3.2.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.12
wrapt==1.14.1
